API introduction, refactoring lexicon codegen.
The codegen code has been restructured for better extensibility in the
future. It also introduces codegen for query parameter types, declaring
a function that allows the user to convert the parameters into a
Map<String, String>.

Also included in here is the beginning of the actual API layer to
communicate with XRPC-based APIs. The only supported method right now is
to create a new session, and the internals of how this method will be
executed and parsed remain to be refactored further possibly. Each call
will return a sealed type to represent the success/failure possibilities
of the request, including deserializing an optional error representation
from the server.