Render a bottom navigation bar.
This necessitates a HomeWorkflow that wraps the existing
TimelineWorkflow, such that the parent AppWorkflow can render the new
home one instead. Home is responsible for coordinating which tab is
selected, as well as rendering full-bleed content *above* the tab bar
and tab content, if necessary.

This seems to have broken remembering the scroll position in the
timeline, though, so that will need to be fixed.