Use custom serializer for sealed value types.
When serializing a value class that is part of a sealed hierarchy, the
inner type's SerialName is used in lieu of the outer type's. This causes
an issue only when serializing data, but it can be mitigated by forcing
the outer type's SerialName to be used. Who knows what new, horrible
unintended side effects this will create, but it seems to fix
serializing things like facet types so let's roll with it for now.