Eradicate any notion of "defs" in generated APIs.
The phrase "defs" currently appears in many places because of its use in
the lexicon schemas, especially as a placeholder of sorts in order to
declare supporting classes used by multiple XRPC requests/responses.
This strips out any reference to them in package or class names so that
they are not present in any generate APIs, save for @SerialName values.