Autogenerate API methods.
For all `query` and `procedure` models, we can autogenerate an API
interface that contains methods accepting the input type (either a
request body, a query parameters list, or neither), and return the
output type (or `Unit` if not specified) contained in a wrapper type
to denote success or failure in invoking that method.

This also generates a class that implements that interface, overriding
each method to invoke the XRPC method on the remote server via a given
Ktor `HttpClient`. The client can be configured in any way that the
developer chooses, and it will have an additional configuration applied
to assist in decoding responses from the XRPC server.