Use value classes to strongly-type string fields.
Certain string fields in lexicons are marked as having a certain type,
such as referencing a DID, a generic URI, a user handle, or so on. In
places where this is defined, the API will use a `value class` that
wraps a `String` corresponding to that specific type. Value classes are
optimal for this approach because they are not assignment-compatible
with one another (unlike `typealias`es), and are nearly as fast to
allocate as the underlying `String` is.