Generate types and methods for subscriptions.
Subscription calls differ from query and procedure calls in that they
are actually subscriptions to a websocket, not a standard HTTP call. For
this reason, they actually respond with a `Flow<*>`, not just one value.

Part of the actual CBOR implementation in kotlinx-serialization is
adapted here to support byte arrays as byte strings for parameterized
types, like List<ByteArray>.