Convert AtIdentifier to sealed interface.
Since this type is either a Did or a Handle, we can make both of those
types inherit from AtIdentifier as a sealed type, and inspect the json
at deserialization time to determine the appropriate concrete type.

It is unambiguous whether an AtIdentifier is a Handle or a Did because a
Did always starts with `did:`, and the colon character (`:`) is not an
allowed in Handles.