Begin implementation of Jetstream service API.
This allows developers to access the firehose in a more friendly way
than via the subscribeRepos() method. In this way, consumers can receive
friendly JSON output instead of CBOR-encoded blocks, as well as easy
filtering on specific collections or DIDs. This also supports the zstd
compression option for JVM and JS which can reduce message sizes
considerably. "A Jetstream consumer that only cares about posts and has
zstd compression enabled can get by on as little as ~25.5GB/mo, <99% of
the full weight firehose."

https://jazco.dev/2024/09/24/jetstream/